+++ tests/apps/good_flow_drop_column_with_constraints/migrations/0003_remove_testtablemain_field_i7.py
@@ -1,0 +1,20 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    atomic = False
+
+    dependencies = [
+        (
+            "good_flow_drop_column_with_constraints",
+            "0002_remove_testtablemain_drop_col_u1_and_more",
+        ),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name="testtablemain",
+            name="field_i7",
+        ),
+    ]