+++ tests/apps/good_flow_drop_column_with_constraints/migrations/0012_remove_testtablemain_field_u5.py
@@ -1,0 +1,20 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    atomic = False
+
+    dependencies = [
+        (
+            "good_flow_drop_column_with_constraints",
+            "0011_remove_testtablemain_field_u6",
+        ),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name="testtablemain",
+            name="field_u5",
+        ),
+    ]