+++ tests/apps/good_flow_drop_column_with_constraints_old/migrations/0006_remove_testtablemain_field_i4.py
@@ -1,0 +1,20 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    atomic = False
+
+    dependencies = [
+        (
+            "good_flow_drop_column_with_constraints_old",
+            "0005_remove_testtablemain_field_i5",
+        ),
+    ]
+
+    operations = [
+        migrations.RemoveField(
+            model_name="testtablemain",
+            name="field_i4",
+        ),
+    ]