+++ CHANGES.md
@@ -4,7 +4,9 @@
   - changed `ADD COLUMN DEFAULT NULL` to safe operation for code default
   - changed `ADD COLUMN DEFAULT NOT NULL` to safe operation for `db_default` in django 5.0+
   - added `ZERO_DOWNTIME_MIGRATIONS_KEEP_DEFAULT` settings and changed `ADD COLUMN DEFAULT NOT NULL` with this settings to safe operation for django<5.0
+  - added `ZERO_DOWNTIME_MIGRATIONS_EXPLICIT_CONSTRAINTS_DROP` settings and added dropping constraints and indexes before drop column or table
   - fixed sqlmigrate in idempotent mode
+  - fixed creation unique constraint with include parameter
   - updated unsafe migrations links to documentation
   - updated patched code to latest django version
   - updated test image to ubuntu 24.04